Update tests for ImportSeed
